Isolate unit tests from filesystem config file
Due to the fact that many modules refer to the config module
it's error prone to mock.patch all of them (specially in api tests
where many sub-modules are loaded) as it can lead to 'leaked' modules
accessing the config file in the filesystem. This patch fixes that by making
the wrapper script that calls the unit tests isolate the modules by setting
the TESSIA_CFG env variable pointing to a temporary file. In addition, change
the handling of config content in unit tests from using mock.patch to using
the same temporary file/env variable approach.

Other changes:
- views.auth: make the authorization decorator determine the login manager
 in run-time (instead of module load time) so that it can be mocked in the unit test.
- api.exceptions: similarly determine the realm from config in runtime (instead of
module load time)