Do not use ignore_ret for reboot command in order to avoid misses
When using ignore_ret parameter sometimes the command was lost and
not executed by the connected host, which led to systems not being
rebooted on installation finish. This is probably due to some kind
of race condition while paramiko flushes the socket. We remedy this
by trying to read the return output after sending the command in
order to assure the command is flushed and thus really executed.
Use a timeout parameter as the reboot will obviously stop the system
and the socket will die anyway.