Adjust cli installation files so that 'pip install' works
Due to the fact that pip copies the directory before calling setup.py
it's not possible to run git commands to find the package version.
The solution is to use a wrapper for pip that calls git and generates
a version file used by setup.py before calling 'pip install'.