Make templates reusable by different OS versions
Rather than linking each template to a specific OS version
they are now independent and can be reused for different OS
versions. The OSes now in turn have a default template
associated so that users can install their systems by
specifying only the OS version while still retaining the
capability to use a custom template if desired.

As the cmdline templates are dependent on the install machine
implementation and should not be changed by users they are
are removed from the db and placed along with the install
machine files.