redhat: activate disks in kickstart's %pre section
The previous approach of using the kernel command line args to specify
the disks to be activated had problems when many entries exist
(specially with FCP paths) due to the limitation on the number of
characters that can be entered. Activating the disks in the %pre
section of the kickstart file solves this problem.