[tessia-cli] Force private key and config files
This patch prevents key and config files from being used if they could
possibly be accessed by others. New files will now be created as read
only for the owner of the file. In addition a warning is shown if
the containing folder is accessible by others. This could otherwise
compromise the security of the tool on systems where the umask is
inappropriately set to share home directories between users.

DCO 1.1 Signed-off-by: André Wild <andre.wild1@ibm.com>