[tessia-cli] changed the prefered output format to table instead of list
This patch changes the default output of the list commands to tables
instead of lists. A new option --long was introduced which shows more
information for some of the commands. The output of the following cmds
has been changed:

- autotemplate list
- conf key-list
- net ip-list
- net subnet-list
- net zone-list
- os list
- perm project-list
- perm user-list
- perm user-roles
- repo list
- storage server-list
- storage server-types
- storage vol-list
- storage vol-types
- system iface-list
- system list
- system model-list
- system prof-list
- system states
- system types

In addition the output of the following commands has been reduced:
- autotemplate list
- net ip-list
- net subnet-list
- net zone-list
- os list
- repo list
- storage server-list
- storage vol-list
- system iface-list
- system list
- system prof-list

With --long the attributes are still shown.

DCO 1.1 Signed-off-by: André Wild <andre.wild1@ibm.com>