Offload permission checks to DB queries
This patch introduces query-time checks for role permissions. Restricted users have had a performance penalty caused by flask first retrieving a list of resources as usual and then checking every one of them against user role and project. This had caused a lot of additional queries to be issued to the database.

Now all 'select' queries are constructed in a way that contains checks for user permissions, so that inaccessible items are never retrieved, and no further client-side processing is required.

Most testcases were not ready for changes in read data and had to be adapted.

Other fixes:
- session rollback was missing on interface attach conflict