Implement a mediator between components to pass secrets without storing them in database.
We'll be using a separate container (redis instance) to store transient data, such as secrets used for starting jobs. This will help us avoid storing sensitive data in the database even briefly.

A redis container is started in db network, so only server has access to it. Compose configuration includes user settings and uri string for redis.

Data volume is mounted to host, and a snapshot is saved by redis to be picked up after a restart. Redis configuration allows it to be completely in-memory - that will cause waiting jobs to fail to start after a service interrupt, because secrets will no longer be available - but it would be the safest possible option.

What constitutes a secret in the parameters or what else can be prepared is left to state machines to decide; by default parmfile is passed unchanged.

BulkOp state machine implements one of the functions, because it requires an additional parameter from request stage.

Ansible state machine parmfile is extended with a secrets dictionary, where key-values (strings only) can be given. The keys can be used in the rest of the parmfile, and will be substituted for the values during execution.