Fix a data race during job request creation
Scheduler runs in a separate process and periodically checks database for new job requests. A data race occurs when a request is committed to  the database, but secrets data is not yet sent to the mediator.

Unfortunately, we need to know the ID of the new request before secrets are passed to mediator, which means we need a two-stage request creation process. It is handled natively by Potion-SQLAlchemy, this patch uses its flush mechanism to have a partial commit within a transaction that should be invisible to scheduler until the full commit.