Allow using a supervising entrypoint for ansible container
Entrypoint in docker containers is responsible for reaping processes that are executed in the namespace of the container, signal propagation and some other things. As such, current solution does not satisfy the requirements and, depending on workload, may lead to zombie processes accumulating over time.

There is no simple answer to which solution could be used as PID 1 process, so this patch only contains changes that are necessary to be able to plug in a different ansible docker image, as long as it is based on the provided Dockerfile.