Remove autoinstall parmfiles that are directories
For Ubuntu Subiquity installer prepares a directory structure that is used by installer. Sometimes it is not cleaned up, and then installation of a different distro on the same profile fails, because it watns to overwrite a file, but target parmfile is a directory.

This patch considers the possibility of a parmfile being a directory