Delay z/VM logon until required
When an installation fails, tessia usually leaves system as it is, which helps with debugging installation issues. However, on z/VM it resets system state by logging in on hypervisor and IPL'ing CMS.

Ths patch delays logon to z/VM until it is required e.g. for booting, so that logon is not performed during cleanup.