Remove the need for duplicate mac address definition in libvirt iface
Currently 'system iface-add' command requires MAC address to be set, and a MACVTAP interface with custom libvirt definition also requires MAC address to be indicated. This caused duplication of data to be specified in the command line, complicated by the fact that libvirt definition is loaded from an external file.

This patch adds MAC address handling for iface-add and iface-edit commands, so that mac argument can be omitted from command line, and in that case its value is taken from libvirt definition.