Implement ContainerSpawner
ContainerSpawner is a new type of spawner for scheduler/looper that runs state machines in their own containers. New containers are based on server image and network connections, but using a different entrypoint (exec.py)

Job data is passed to the spawned container over stdin as a serialized json. For robustness a streaming JSON parser is used to read the jobs, instead of waiting for the stdin to close.

With containerized state machines an issue was uncovered in ansible state machine implementation, where ownership on the user directory in ansible runner container was overwritten, thus interfering with the execution. This was incidentally not detected earlier, because user id 1000 is used as the current user in both tessia server and ansible image and so no overwrite is seen.

In order for autoinstall state machines to work, static web directory had to be additionally mapped into tessia server container to be further remapped in executors. Compose definition was updated to provide additional environment variables that define tessia server instance.

With this patch ansible jobs survive compose service restart.