Do not fail poweron job on verification error
A vexing failure of poweron job caused by verification mismatch is now opt-in with a CLI flag "--verify", with default being no verification.

It is more common that not that system state is slightly off compared to expected in the tessia database; such discrepancies are still logged in job output, but do not produce an error unless a verification flag is set.

DCO 1.1 Signed-off-by: Alexander Efremkin <aefr@de.ibm.com>