tela/Dockerfile: Configure SSH-Client for early reboot/crash detection
When a kernel panic occurs during the execution of a testcase
via tessia the affected system will take a kdump and reboot.
In the meantime the tessia server keeps waiting on the ssh
connection, which times out after 2 hours (TCP timeout,
can be checked via 'sysctl -ar ".*tcp_keepalive_time"').
After the timeout the test execution fails, as expected.
When this occurs in the automatic test environment, all
subsequent test cases will be delayed accordingly.

To reduce the waiting time, set ServerAliveInterval and
ServerAliveCountMax in the ssh client config file to
3x360 seconds wait time. This allows the tela state machine
to detect crashed or rebooted systems in ~18 minutes.

And add the parameter -e to the echo call to translate the
backslash characters correctly.

Signed-off-by: Karsten Graul <kgraul@linux.ibm.com>